# Make compilation 2x faster by using literals