<!--- Provide a general summary of your changes in the Title above -->

## Description

I replaced `Character(Unicode.Scalar(...)!)` by `"\u{...}"`. This makes compilation almost 2x faster.

Here is the benchmark:

| branch | command | 1st execution time | 2nd execution time | 3rd execution time | average |
|:------:|:--------|-------------------:|-------------------:|-------------------:|-------:|
| master | `swift build` | 4.37s | 3.65s | 3.52s | 3.85s |
| replace-by-literal | `swift build` | 2.11s | 2.17s | 2.11s | 2.13s |
| master | `swift build -c release` | 9.81s | 9.37s | 9.67s | 9.62s |
| replace-by-literal | `swift build -c release` | 5.04s | 5.11s | 5.02s | 5.06s |

(Mac mini (M1, 2020) 8GB RAM, macOS 12.3, Xcode 13.3, Swift 5.6)

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

My changes can reduce compile time without changing any functionality.

Not only that, these can also resolve issues like https://github.com/kkk669/DevToys.swiftpm/pull/6#issuecomment-1069332639.

## How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] I have submitted a [CLA form](https://github.com/IBM-Swift/CLA)
- [x] If applicable, I have updated the documentation accordingly.
- [x] If applicable, I have added tests to cover my changes.