@youming-lin 2022-03-17
This looks like a complete revert of #51 minus changes in two lines of comments.
LGTM since I originally wrote this using literals, but according to the thread in #51 the bug in Swift wasn't fixed until 5.1 so this change will break for users on 5.0.

I don't work on this project any more so I'll defer to @dannys42 to decide if 5.0 support is still needed.