@dannys42 2022-03-18
Thanks @kkk669 for the submission and @youming-lin for the context.  I've been slowly phasing out Swift 5.0 and 5.1 from the various Kitura projects.  My typical approach has been to bump the major version of the project while simultaneously dropping support for older versions of Swift and updating the CI build.  So let me take a stab at that then revisit this.

I'll try to address this as soon as I can.  But I apologize right now if I don't get to this soon, I'm a bit overloaded right now.  Feel free to ping a reminder to me if I'm taking too long.