@kkk669 2022-03-18
@youming-lin Thank you for sharing the related issue. I also replaced `Character(Scalar(...)!)` in the comments. 6d490b02c2b376d95e23b00fa4ce754532046980

@dannys42 Thank you for sharing the future plan. I'm going to wait until then.