@kkk669 2022-03-25
I think the values should be left as they are even if it is an incorrect value. By doing so, this PR becomes a simple conversion-only difference.

If `\u{338}` is incorrect, we should correct it in another PR.