@youming-lin 2022-03-25
I originally generated the maps from a JSON file I found that maps the named character references to unicodes, e.g. https://html.spec.whatwg.org/entities.json

I would doublecheck the [spec](https://html.spec.whatwg.org/multipage/named-characters.html) to make sure the unicodes are correct.