@kkk669 2022-03-25
@youming-lin Thank you for sharing the specs. I checked that and I confirmed that #51's change was incorrect. However, I haven't made any changes to that part in this PR, so if I fix it here, I think it would be out of line with the purpose of this PR.