@dannys42 2022-03-29
@kkk669 I pushed your changes and made a release for 4.0.1.  I also added an [issue](https://github.com/Kitura/swift-html-entities/issues/67) to track the unicode bug.  Please feel free to add any context you think is necessary for that issue.  Or submit a new PR if you're inclined. :-)

There's an issue with cocoapods so I can't push an updated pod right now.  I'll try again later.